Move most of geckolib into style::gecko
