+++ components/style/Cargo.toml
@@ -13,7 +13,7 @@ path = "lib.rs"
 
 [features]
 gecko = []
-servo = ["serde/unstable", "serde_macros", "heapsize_plugin",
+servo = ["serde/unstable", "serde", "serde_macros", "heapsize_plugin",
          "style_traits/servo", "app_units/plugins",
          "cssparser/heap_size", "cssparser/serde-serialization",
          "selectors/heap_size", "selectors/unstable", "string_cache",
@@ -33,15 +33,17 @@ heapsize = "0.3.0"
 heapsize_plugin = {version = "0.1.2", optional = true}
 lazy_static = "0.2"
 log = "0.3.5"
+libc = "0.2"
 matches = "0.1"
 num-integer = "0.1.32"
 num-traits = "0.1.32"
+num_cpus = "0.2.2"
 ordered-float = "0.2.2"
 quickersort = "2.0.0"
 rand = "0.3"
 rustc-serialize = "0.3"
 selectors = "0.13"
-serde = "0.8"
+serde = {version = "0.8", optional = true}
 serde_macros = {version = "0.8", optional = true}
 smallvec = "0.1"
 string_cache = {version = "0.2.26", features = ["heap_size"], optional = true}
