+++ components/style/binding_tools/regen_atoms.py
@@ -138,7 +138,8 @@ def write_atom_macro(atoms, file_name):
         f.write("}\n\n")
         f.write("#[macro_export]\n")
         f.write("macro_rules! atom {\n")
-        f.writelines(['("%s") => { $crate::string_cache::atom_macro::unsafe_atom_from_static($crate::string_cache::atom_macro::%s as *mut _) };\n'
+        f.writelines(['("%s") => { $crate::string_cache::atom_macro::unsafe_atom_from_static(\
+                      $crate::string_cache::atom_macro::%s as *mut _) };\n'
                      % (atom.value, atom.ident) for atom in atoms])
         f.write("}\n")
 
@@ -192,5 +193,5 @@ def write_pseudo_element_helper(atoms, target_filename):
 def build(objdir, verbose=False):
     atoms = collect_atoms(objdir)
     write_atom_macro(atoms, "../gecko_string_cache/atom_macro.rs")
-    write_pseudo_element_helper(atoms, "../generated/gecko_pseudo_element_helper.rs")
+    write_pseudo_element_helper(atoms, "../gecko/generated/gecko_pseudo_element_helper.rs")
     return 0
