+++ components/style/gecko/context.rs
@@ -2,9 +2,9 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
+use context::{LocalStyleContext, StyleContext, SharedStyleContext};
 use std::cell::RefCell;
 use std::rc::Rc;
-use style::context::{LocalStyleContext, StyleContext, SharedStyleContext};
 
 thread_local!(static LOCAL_CONTEXT_KEY: RefCell<Option<Rc<LocalStyleContext>>> = RefCell::new(None));
