+++ components/style/gecko/wrapper.rs
@@ -4,51 +4,52 @@
 
 #![allow(unsafe_code)]
 
-use style::gecko_bindings::bindings;
-use style::gecko_bindings::bindings::{Gecko_CalcStyleDifference, Gecko_StoreStyleDifference};
-use style::gecko_bindings::bindings::{Gecko_DropStyleChildrenIterator, Gecko_MaybeCreateStyleChildrenIterator};
-use style::gecko_bindings::bindings::{Gecko_ElementState, Gecko_GetDocumentElement};
-use style::gecko_bindings::bindings::{Gecko_GetFirstChild, Gecko_GetFirstChildElement};
-use style::gecko_bindings::bindings::{Gecko_GetLastChild, Gecko_GetLastChildElement};
-use style::gecko_bindings::bindings::{Gecko_GetNextSibling, Gecko_GetNextSiblingElement, Gecko_GetNextStyleChild};
-use style::gecko_bindings::bindings::{Gecko_GetNodeFlags, Gecko_SetNodeFlags, Gecko_UnsetNodeFlags};
-use style::gecko_bindings::bindings::{Gecko_GetParentElement, Gecko_GetParentNode};
-use style::gecko_bindings::bindings::{Gecko_GetPrevSibling, Gecko_GetPrevSiblingElement};
-use style::gecko_bindings::bindings::{Gecko_GetServoDeclarationBlock, Gecko_IsHTMLElementInHTMLDocument};
-use style::gecko_bindings::bindings::{Gecko_IsLink, Gecko_IsRootElement, Gecko_IsTextNode};
-use style::gecko_bindings::bindings::{Gecko_IsUnvisitedLink, Gecko_IsVisitedLink};
-use style::gecko_bindings::bindings::{Gecko_LocalName, Gecko_Namespace, Gecko_NodeIsElement};
-use style::gecko_bindings::bindings::Gecko_ClassOrClassList;
-use style::gecko_bindings::bindings::Gecko_GetStyleContext;
-use style::gecko_bindings::structs::{NODE_HAS_DIRTY_DESCENDANTS_FOR_SERVO, NODE_IS_DIRTY_FOR_SERVO};
-use style::gecko_bindings::structs::{RawGeckoDocument, RawGeckoElement, RawGeckoNode};
-use style::gecko_bindings::structs::{nsChangeHint, nsIAtom, nsStyleContext};
-use style::gecko_bindings::structs::OpaqueStyleData;
-use style::gecko_bindings::sugar::ownership::FFIArcHelpers;
-use style::gecko_string_cache::{Atom, Namespace, WeakAtom, WeakNamespace};
-use glue::GeckoDeclarationBlock;
+
+use data::PrivateStyleData;
+use dom::{LayoutIterator, NodeInfo, TDocument, TElement, TNode, TRestyleDamage, UnsafeNode};
+use dom::{OpaqueNode, PresentationalHintsSynthetizer};
+use element_state::ElementState;
+use error_reporting::StdoutErrorReporter;
+use gecko::selector_impl::{GeckoSelectorImpl, NonTSPseudoClass, PseudoElement};
+use gecko::snapshot::GeckoElementSnapshot;
+use gecko::snapshot_helpers;
+use gecko_bindings::bindings;
+use gecko_bindings::bindings::{Gecko_CalcStyleDifference, Gecko_StoreStyleDifference};
+use gecko_bindings::bindings::{Gecko_DropStyleChildrenIterator, Gecko_MaybeCreateStyleChildrenIterator};
+use gecko_bindings::bindings::{Gecko_ElementState, Gecko_GetDocumentElement};
+use gecko_bindings::bindings::{Gecko_GetFirstChild, Gecko_GetFirstChildElement};
+use gecko_bindings::bindings::{Gecko_GetLastChild, Gecko_GetLastChildElement};
+use gecko_bindings::bindings::{Gecko_GetNextSibling, Gecko_GetNextSiblingElement, Gecko_GetNextStyleChild};
+use gecko_bindings::bindings::{Gecko_GetNodeFlags, Gecko_SetNodeFlags, Gecko_UnsetNodeFlags};
+use gecko_bindings::bindings::{Gecko_GetParentElement, Gecko_GetParentNode};
+use gecko_bindings::bindings::{Gecko_GetPrevSibling, Gecko_GetPrevSiblingElement};
+use gecko_bindings::bindings::{Gecko_GetServoDeclarationBlock, Gecko_IsHTMLElementInHTMLDocument};
+use gecko_bindings::bindings::{Gecko_IsLink, Gecko_IsRootElement, Gecko_IsTextNode};
+use gecko_bindings::bindings::{Gecko_IsUnvisitedLink, Gecko_IsVisitedLink};
+use gecko_bindings::bindings::{Gecko_LocalName, Gecko_Namespace, Gecko_NodeIsElement};
+use gecko_bindings::bindings::Gecko_ClassOrClassList;
+use gecko_bindings::bindings::Gecko_GetStyleContext;
+use gecko_bindings::structs::{NODE_HAS_DIRTY_DESCENDANTS_FOR_SERVO, NODE_IS_DIRTY_FOR_SERVO};
+use gecko_bindings::structs::{RawGeckoDocument, RawGeckoElement, RawGeckoNode};
+use gecko_bindings::structs::{nsChangeHint, nsIAtom, nsStyleContext};
+use gecko_bindings::structs::OpaqueStyleData;
+use gecko_bindings::sugar::ownership::{FFIArcHelpers, HasArcFFI, HasFFI};
 use libc::uintptr_t;
+use parser::ParserContextExtraData;
+use properties::{ComputedValues, parse_style_attribute};
+use properties::PropertyDeclarationBlock;
+use refcell::{Ref, RefCell, RefMut};
+use selector_impl::ElementExt;
+use selector_matching::ApplicableDeclarationBlock;
 use selectors::Element;
 use selectors::parser::{AttrSelector, NamespaceConstraint};
-use snapshot::GeckoElementSnapshot;
-use snapshot_helpers;
+use sink::Push;
 use std::fmt;
 use std::ops::BitOr;
 use std::ptr;
 use std::sync::Arc;
-use style::data::PrivateStyleData;
-use style::dom::{LayoutIterator, NodeInfo, TDocument, TElement, TNode, TRestyleDamage, UnsafeNode};
-use style::dom::{OpaqueNode, PresentationalHintsSynthetizer};
-use style::element_state::ElementState;
-use style::error_reporting::StdoutErrorReporter;
-use style::parser::ParserContextExtraData;
-use style::properties::{ComputedValues, parse_style_attribute};
-use style::properties::PropertyDeclarationBlock;
-use style::refcell::{Ref, RefCell, RefMut};
-use style::selector_impl::ElementExt;
-use style::selector_matching::ApplicableDeclarationBlock;
-use style::sink::Push;
-use style::gecko::selector_impl::{GeckoSelectorImpl, NonTSPseudoClass, PseudoElement};
+use std::sync::atomic::{AtomicBool, AtomicPtr};
+use string_cache::{Atom, Namespace, WeakAtom, WeakNamespace};
 use url::Url;
 
 pub struct NonOpaqueStyleData(RefCell<PrivateStyleData>);
@@ -58,6 +59,24 @@ impl NonOpaqueStyleData {
         NonOpaqueStyleData(RefCell::new(PrivateStyleData::new()))
     }
 }
+
+
+pub struct GeckoDeclarationBlock {
+    pub declarations: Option<Arc<PropertyDeclarationBlock>>,
+    // XXX The following two fields are made atomic to work around the
+    // ownership system so that they can be changed inside a shared
+    // instance. It wouldn't provide safety as Rust usually promises,
+    // but it is fine as far as we only access them in a single thread.
+    // If we need to access them in different threads, we would need
+    // to redesign how it works with MiscContainer in Gecko side.
+    pub cache: AtomicPtr<bindings::nsHTMLCSSStyleSheet>,
+    pub immutable: AtomicBool,
+}
+
+unsafe impl HasFFI for GeckoDeclarationBlock {
+    type FFIType = bindings::ServoDeclarationBlock;
+}
+unsafe impl HasArcFFI for GeckoDeclarationBlock {}
 
 
 // We can eliminate OpaqueStyleData when the bindings move into the style crate.
