+++ components/style/gecko_string_cache/namespace.rs
@@ -2,12 +2,12 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-use string_cache::{Atom, WeakAtom};
 use gecko_bindings::structs::nsIAtom;
 use selectors::bloom::BloomHash;
 use std::borrow::Borrow;
 use std::fmt;
 use std::ops::Deref;
+use string_cache::{Atom, WeakAtom};
 
 #[macro_export]
 macro_rules! ns {
